+++ goshop-service-order/src/main/java/org/goshop/order/mapper/master/GsStoreCartMapper.java
@@ -5,9 +5,9 @@ import org.goshop.order.pojo.GsStoreCart;
 public interface GsStoreCartMapper {
     int deleteByPrimaryKey(Long id);
 
-    int insert(GsStoreCart record);
+    long insert(GsStoreCart record);
 
-    int insertSelective(GsStoreCart record);
+    long insertSelective(GsStoreCart record);
 
     int updateByPrimaryKeySelective(GsStoreCart record);
